test: add test for workspace-folder-level settings
In the past, ShellCheck has supported a single set of preferences per
workspace, assuming that those preferences should always hold for the
entire workspace. However, VS Code supports settings on more levels,
for example workspace-folder-level and language-specific settings. [1]

This test creates a workspace with two folders, each with independent
settings, and asserts that for each folder, the ShellCheck extension
honors the folder-specific settings.

[1]: https://code.visualstudio.com/docs/getstarted/settings#_settings-precedence