ci: make compatible with act
Act [1] is a tool that allows you to run GitHub Actions locally.

Because of the unique way in which GitHub’s environment for Actions is
configured [2], many GitHub Actions use packages that a standard Ubuntu
image doesn’t provide, and assume those packages to be present. This
usually causes GitHub Action workflows to not work out of the box with
`act`.

In order to enable local debugging with `act`, add the following missing
packages to the container image in `ci.yaml`:

- `libasound`
- `libgbm`
- `libgtk`
- `libnss3`
- `sudo`

To debug vscode-shellcheck’s CI locally, run the following command line:

```
sudo act -j test -P ubuntu-latest=node:16-buster
```

The first four packages in the list are dependencies of Electron or
VS Code. The `sudo` package is required by xvfb-action.

[1]: https://github.com/nektos/act

[2]: https://github.com/nektos/act#default-runners-are-intentionally-incomplete